orchestration/pyproject.toml
[tool.poetry]
name = "hca_orchestration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.9"
argo-workflows = "^5.0.0"
cached-property = "^1.5.2"
dagster = "0.12.10"
dagster-k8s = "0.12.10"
dagster-postgres = "0.12.10"
dagster-slack = "0.12.10"
data-repo-client = "^1.134.0"
google-cloud-bigquery = "^2.4.0"
google-cloud-storage = "^1.3.5"
python-dateutil = "^2.8.1"
typing-extensions = "^3.7.4"
pyyaml = "^5.3"
dagster-gcp = "^0.12.10"
broad-dagster-utils = "0.6.4"
hca-import-validation = "^0.0.1"
more-itertools = "^8.8.0"
graphql-ws = "<0.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
dagit = "0.12.10"
flake8 = "^3.8.4"
mypy = "^0.812"
pdbpp = "^0.10.2"
pre-commit = "^2.11.0"
pytest = "^6.2.1"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
check = "hca_manage.check:run"
dataset = "hca_manage.dataset:run"
snapshot = "hca_manage.snapshot:run"
soft_delete = "hca_manage.soft_delete:run"

[build-system]
requires = ["poetry-core=^1.1.8"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
aggressive = 1
exclude = ".pytest_cache,__pycache__"
max_line_length = 120
